用户和管理员的登录注册，以及权限管理(token)

+++ src/main/java/com/twinkle/service/LabelService.java
@@ -3,9 +3,13 @@ package com.twinkle.service;
 import com.twinkle.dao.LabelDao;
 import com.twinkle.pojo.Label;
 import com.twinkle.util.IdWorker;
+import com.twinkle.util.JwtUtil;
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpRequest;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
 import java.util.List;
 
@@ -19,23 +23,36 @@ public class LabelService {
     @Autowired
     private IdWorker idWorker;
 
+    //通过request获取请求偷
+    @Autowired
+    private HttpServletRequest httpRequest;
+
+    @Autowired
+    private JwtUtil jwtUtil;
+
     /**
      * 保存分类
+     *
      * @param label
      */
     public void save(Label label) {
-        label.setLabelId(idWorker.nextId()+"");
+        label.setLabelId(idWorker.nextId() + "");
         label.setLabelName(label.getLabelName().toLowerCase());
         labelDao.save(label);
     }
 
 
-
     /**
      * 删除分类
+     *
      * @param labelName
      */
     public void delete(String labelName) {
+        String admin_claims = (String) httpRequest.getAttribute("admin_claims");
+        System.out.println(admin_claims);
+        if(admin_claims.isEmpty()){
+            throw new RuntimeException("权限不足");
+        }
         labelDao.deleteByLabelName(labelName);
     }
 
@@ -50,12 +67,13 @@ public class LabelService {
     /**
      * 根据id查询
      */
-    public Label findByLabelId(String labelId){
+    public Label findByLabelId(String labelId) {
         return labelDao.findByLabelId(labelId);
     }
 
     /**
      * 根据分类名查询
+     *
      * @param labelName
      * @return
      */
